F Cancels requested loading image task
